Add https action helper
The action helper is a small class to be able to control which
pages are explicitly served in http and which in https. In action
controllers, it is easy to mark several actions to be forced in
https or http mode.

See phpdoc for more information and a usage example.